+++ R/data_arrange.R
@@ -24,7 +24,6 @@
 #' data_arrange(head(mtcars), c("gear", "foo"), safe = FALSE)
 #' }
 #' @export
-
 data_arrange <- function(data, select = NULL, safe = TRUE) {
   UseMethod("data_arrange")
 }
@@ -32,7 +31,6 @@ data_arrange <- function(data, select = NULL, safe = TRUE) {
 
 
 #' @export
-
 data_arrange.default <- function(data, select = NULL, safe = TRUE) {
   if (is.null(select) || length(select) == 0) {
     return(data)
@@ -106,7 +104,6 @@ data_arrange.default <- function(data, select = NULL, safe = TRUE) {
 
 
 #' @export
-
 data_arrange.grouped_df <- function(data, select = NULL, safe = TRUE) {
   # works only for dplyr >= 0.8.0
   grps <- attr(data, "groups", exact = TRUE)